Commands are now subcommands
Changed commands to be subcommands.
However, we still need a name for the feature. This name should then also be used in the web api which entails refactoring a bit.
